requirements.txt
mysql-connector-python>=8.0.0
click>=8.0.0
configparser>=5.0.0
tabulate>=0.8.0
colorama>=0.4.4
tqdm>=4.62.0
pyinstaller>=5.0.0
pytest>=7.0.0
typing-extensions>=4.0.0
python-dotenv>=0.20.0
